FIX keep 16bit precision for metal pbr

+++ src/render/FrameBufferAttachment.cpp
@@ -219,7 +219,7 @@ FrameBufferAttachment FrameBufferAttachment::getGBufferMetal(GLenum attachment)
 {
     FrameBufferAttachment spec;
     spec.type = FrameBufferAttachmentType::texture;
-    spec.internalFormat = GL_RGBA8;
+    spec.internalFormat = GL_RGBA16F;
     spec.attachment = attachment;
     spec.useDrawBuffer = true;
     spec.clearMask = GL_COLOR_BUFFER_BIT;
